+++ src/Symfony/Component/Security/Core/Authentication/Token/TokenInterface.php
@@ -41,23 +41,22 @@ interface TokenInterface
      * Returns the user credentials.
      *
      * @return mixed The user credentials
+     *
+     * @deprecated since 5.4
      */
     public function getCredentials();
 
     /**
      * Returns a user representation.
      *
-     * @return string|\Stringable|UserInterface
+     * @return UserInterface
      *
      * @see AbstractToken::setUser()
      */
     public function getUser();
 
     /**
-     * Sets the user in the token.
-     *
-     * The user can be a UserInterface instance, or an object implementing
-     * a __toString method or the username as a regular string.
+     * Sets the authenticated user in the token.
      *
      * @throws \InvalidArgumentException
      */